make it easier to test and read
a new class, XML.pm, will be the adaptor which generates XML requests
making it easier to test the XML creation on its own.

Base.pm (and/or NNCIPP.pm) will rely on it to generate (and possibly parse)
NCIP messages.

XML should not be completely independent from Koha or any configuration.